tasks/11_ml_new_model/check_function/requirements.txt
functions-framework==3.8.2
Flask==2.3.2
google-cloud-storage
pandas==1.5.3
numpy==1.24.4
geopandas==0.13.2
Shapely==2.1.0
scikit-learn==1.3.0
geopy==2.4.4